Generate vault endpoints

+++ src/Opdex.Client/Model/CompleteVaultProposalEvent.cs
@@ -39,7 +39,7 @@ namespace Opdex.Client.Model
         /// <param name="eventType">eventType.</param>
         /// <param name="contract">An address on the Cirrus network.</param>
         /// <param name="sortOrder">Index to sort event order.</param>
-        public CompleteVaultProposalEvent(decimal proposalId = default(decimal), bool approved = default(bool), TransactionEventType eventType = default(TransactionEventType), string contract = default(string), int sortOrder = default(int))
+        public CompleteVaultProposalEvent(int proposalId = default(int), bool approved = default(bool), TransactionEventType eventType = default(TransactionEventType), string contract = default(string), int sortOrder = default(int))
         {
             this.ProposalId = proposalId;
             this.Approved = approved;
@@ -53,7 +53,7 @@ namespace Opdex.Client.Model
         /// </summary>
         /// <value>Identifier for the proposal</value>
         [DataMember(Name = "proposalId", EmitDefaultValue = false)]
-        public decimal ProposalId { get; set; }
+        public int ProposalId { get; set; }
 
         /// <summary>
         /// Whether the proposal was approved
@@ -185,8 +185,8 @@ namespace Opdex.Client.Model
         /// <returns>Validation Result</returns>
         public IEnumerable<System.ComponentModel.DataAnnotations.ValidationResult> Validate(ValidationContext validationContext)
         {
-            // ProposalId (decimal) minimum
-            if (this.ProposalId < (decimal)1)
+            // ProposalId (int) minimum
+            if (this.ProposalId < (int)1)
             {
                 yield return new System.ComponentModel.DataAnnotations.ValidationResult("Invalid value for ProposalId, must be a value greater than or equal to 1.", new [] { "ProposalId" });
             }
